修改并测试了block，发现并修复了virtualdisk中的若干bug，调整tests目录结构

+++ virtual_disk.h
@@ -8,12 +8,14 @@
  * 对虚拟的磁盘文件进行读写等操作
  * 以扇区为操作单位，每个扇区512个字节
  *
+ * 扇区号32位，最大支持2TB的硬盘
+ *
  */
 
 #define SECTOR_SIZE 512
 #define VDISK_ERROR -1
 
-typedef int32_t vdisk_handle_t; /* 虚拟磁盘句柄 */
+typedef int16_t vdisk_handle_t; /* 虚拟磁盘句柄 */
 
 /* 将一个文件添加为虚拟磁盘 */
 vdisk_handle_t vdisk_add(const char* filepath);
@@ -22,11 +24,11 @@ vdisk_handle_t vdisk_add(const char* filepath);
 int vdisk_remove(vdisk_handle_t handle);
 
 /* 从虚拟磁盘读入数据（以扇区为单位） */
-uint64_t vdisk_read(vdisk_handle_t handle, uint64_t sector, uint64_t count,
-                    char* buf);
+int64_t vdisk_read(vdisk_handle_t handle, uint32_t sector, uint32_t count,
+                   char* buf);
 
 /* 向虚拟磁盘写入数据（以扇区为单位） */
-uint64_t vdisk_write(vdisk_handle_t handle, uint64_t sector, uint64_t count,
-                     const char* buf);
+int64_t vdisk_write(vdisk_handle_t handle, uint32_t sector, uint32_t count,
+                    const char* buf);
 
 #endif
